+++ test-data/test-files/paredit_sandbox.clj
@@ -109,6 +109,25 @@ string. "
 ; Should not delete `#`
 (#|())
 
+;Should delete entire #() form with backspace and delete
+#()
+
+; Should not delete #  
+#{|21 2}
+
+; Should delete #
+#|{21}
+
+;; Backspace should delete the #
+[|#(prn "hello")]
+|#(prn "hello"|)
+
+;; Delete should delete space and #
+[# |(prn "hello")]
+
+; Delete should move before # without deleting it
+(#(|inc 2))
+
 ;; Pair selecting
 
 (cond